fix: refactor on incorporations marketplace
