+++ src/renderer/marketplace/incorporation/checkout/payment-confirmation-content.jsx
