+++ src/renderer/marketplace/incorporation/checkout/payment-confirmation.jsx
