+++ src/renderer/marketplace/incorporation/common/index.js
@@ -1,14 +1,6 @@
 import TreatiesMap from './treaties-map';
 import TreatiesTable from './treaties-table';
-import CountryInfo from './country-info';
 import IncorporationsKYC from './kyc-requirements';
 import { getIncorporationPrice, getTemplateID } from './helpers';
 
-export {
-	TreatiesMap,
-	TreatiesTable,
-	CountryInfo,
-	IncorporationsKYC,
-	getIncorporationPrice,
-	getTemplateID
-};
+export { TreatiesMap, TreatiesTable, IncorporationsKYC, getIncorporationPrice, getTemplateID };
