+++ src/renderer/marketplace/incorporation/details/description-tab.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import { withStyles } from '@material-ui/core/styles';
+import { sanitize } from '../../common';
+
+const styles = theme => ({});
+
+export const DescriptionTab = withStyles(styles)(({ program }) => {
+	return (
+		<div
+			dangerouslySetInnerHTML={{
+				__html: sanitize(program.translation.introduction)
+			}}
+		/>
+	);
+});
+
+export default DescriptionTab;
