+++ src/renderer/marketplace/incorporation/index.jsx
@@ -1,13 +1,10 @@
 import React, { Component } from 'react';
 import { Route } from 'react-router-dom';
-
-import IncorporationsTable from './table';
-import IncorporationsDetailView from './detail';
-import {
-	IncorporationCheckout,
-	IncorporationProcessStarted,
-	IncorporationPaymentConfirmation
-} from './pay';
+import { IncorporationTableContainer } from './list/offers-container';
+import IncorporationsDetailsContainer from './details/details-container';
+import IncorporationCheckoutContainer from './checkout/checkout-container';
+import IncorporationProcessStartedContainer from './process-started/process-started-container';
+import IncorporationPaymentConfirmation from './checkout/payment-confirmation';
 
 class MarketplaceIncorporationComponent extends Component {
 	render() {
@@ -15,14 +12,14 @@ class MarketplaceIncorporationComponent extends Component {
 
 		return (
 			<div>
-				<Route exact path={`${path}`} component={IncorporationsTable} />
+				<Route exact path={`${path}`} component={IncorporationTableContainer} />
 				<Route
 					path={`${path}/details/:companyCode/:countryCode/:templateId?`}
-					component={IncorporationsDetailView}
+					component={IncorporationsDetailsContainer}
 				/>
 				<Route
 					path={`${path}/pay/:companyCode/:countryCode/:templateId?`}
-					component={IncorporationCheckout}
+					component={IncorporationCheckoutContainer}
 				/>
 				<Route
 					path={`${path}/pay-confirmation/:companyCode/:countryCode/:templateId?/:confirmation?`}
@@ -30,7 +27,7 @@ class MarketplaceIncorporationComponent extends Component {
 				/>
 				<Route
 					path={`${path}/process-started/:companyCode/:countryCode/:templateId?`}
-					component={IncorporationProcessStarted}
+					component={IncorporationProcessStartedContainer}
 				/>
 			</div>
 		);
