--- src/renderer/marketplace/incorporation/pay/index.js
@@ -1,5 +1,0 @@
-import IncorporationCheckout from './checkout';
-import IncorporationProcessStarted from './process-started';
-import IncorporationPaymentConfirmation from './payment-confirmation';
-
-export { IncorporationCheckout, IncorporationProcessStarted, IncorporationPaymentConfirmation };
